requirements.txt
py_cui
requests
setuptools
python-dotenv>=1.0.0